feat!: remove the one-shot tandem switch command
Ctrl-] is the only flip. _report_switch moves to flip.py (its only
caller); the _SESSION_ID prompt-dispatch thread-through is dead code
without the prompt and goes with it.
